// Angular/src/app/model/lookups.ts
export class csLookups{
      ClientTypeID: number;
      ClientTypeDescription: string ;
      UserModified: string ;
      LastmodifiedDate
      Active:boolean
}

export class csRank{
    RankID: number;
    RankDescription: string ;
    UserModified: string ;
    LastmodifiedDate
    Active:boolean
}

export class csTitle{
    TitleID: number;
    TitleDescription: string ;
    UserModified: string ;
    LastmodifiedDate
    Active:boolean
}

export class csNationality{
    NationalityID: number;
    NationalityDescription: string ;
    UserModified: string ;
    LastmodifiedDate
    Active:boolean
}

export class dipRefundRate{
    RefundRateID: number;
    RefundTypeDescription: string ;
    RefundFuelCode: string ;
    RefundExciseCode: string ;
    RefundExciseDuty: number ;
    RefundFuelLevy: number ;
    UserModified: string ;
    LastmodifiedDate: Date;
    RefundEffectiveDate: Date;
}